requirements.txt
# --- Discord & Network ---
discord.py==2.7.1
aiohttp==3.13.5
requests==2.33.1
pytest==9.0.3

# --- Security & Environment ---
python-dotenv==1.2.2
aiosqlite==0.22.1
sentry-sdk==2.57.0

# --- AI & Machine Learning (Inference Only) ---
torch==2.11.0
safetensors==0.7.0
scikit-learn==1.8.0
skops==0.13.0
matplotlib==3.10.8
pandas==3.0.2
aiofiles==25.1.0